+++ libraries/psibase/common/src/crypto.cpp
@@ -6,7 +6,7 @@
 
 namespace psibase
 {
-   eosio::checksum256 sha256(const char* data, uint32_t length)
+   eosio::checksum256 sha256(const char* data, size_t length)
    {
       std::array<unsigned char, 256 / 8> result;
       SHA256((const unsigned char*)data, length, result.data());
